bugfix: flyway 오류 수정

+++ src/main/resources/db/migration/V202601110224__add_user_activity_metrics.sql
@@ -8,4 +8,4 @@ ALTER TABLE users
 
 -- 3. 응답률 (소수점 지원)
 ALTER TABLE users
-    ADD COLUMN reply_rate DOUBLE DEFAULT 0.0;
+    ADD COLUMN reply_rate DOUBLE PRECISION DEFAULT 0.0;
